piece.py
from settings import GREEN, BLACK, WHITE


class PIECE:
    COLOR_SET = {GREEN, BLACK, WHITE}

    def __init__(self):
        self.color = GREEN

    def set_color(self, color):
        if color in self.COLOR_SET:
            self.color = color
        else:
            raise Exception("設定可能値以外")

    def turn_color(self):
        if self.color == GREEN:
            raise Exception("反転不可")
        self.color = WHITE if self.color == BLACK else BLACK

    def get_color(self):
        return self.color
